+++ R/use_index_rmd.R
@@ -1,11 +1,6 @@
 #' Use index.Rmd (used for pkgdown::build_home())
-#' @param pkg package description in list format (default: \cr
-#' pkg = list(name = "kwb.umberto",\cr
-#' title = "R package supporting UMERTO LCA at KWB",\cr
-#' desc = "Helper functions for data aggregation and visualisation of\cr
-#' UMBERTO (https://www.ifu.com/umberto) model output"))\cr
 #' @param user user name or organisation under which repository defined in\cr
-#' parameter "repo" is hosted (default: "KWB-R")
+#' parameter "repo" is hosted (default: "KWB-R")\cr
 #' @param domain under which repository is hosted (default: "github")
 #' @param stage badge declares the developmental stage of a package, according
 #' to [https://www.tidyverse.org/lifecycle/](https://www.tidyverse.org/lifecycle/),
@@ -13,14 +8,15 @@
 #' "archived", "dormant", "questioning"), (default: "experiment")
 #' @return generates travis badge link
 #' @export
+#' @importFrom desc desc
+
 use_index_rmd <- function(
-  pkg = list(name = "kwb.umberto",
-desc = paste("Helper functions for data aggregation and",
-"visualisation of UMBERTO (https://www.ifu.com/umberto)
-model output")),
   user = "KWB-R",
   domain = "github",
   stage = "experimental") {
+
+  pkg <- read_description()
+
 
   index_rmd <- c(use_badge_appveyor(pkg$name, user,domain),
                  use_badge_travis(pkg$name, user),
@@ -30,7 +26,7 @@ model output")),
                  "",
                  pkg$desc,
                  "",
-                 use_installation(pkg$name, user, domain))
+                 use_installation(pkg$name, user, domain,output = "rmd"))
 
 
   writeLines(index_rmd, "index.Rmd")