+++ R/use_installation.R
@@ -3,11 +3,15 @@
 #' @param user user name or organisation under which repository defined in
 #' parameter "repo" is hosted
 #' @param domain under which repository is hosted
-use_installation <- function(pkgname, user, domain) {
+#' @param output  if "rmd" (other code chunks are used); otherwise: "md" style!
+#' (default: "md")
+use_installation <- function(pkgname, user, domain, output = "md") {
 
   c("# Installation",
     "",
-    "```{r echo = TRUE, eval = FALSE}",
+    ifelse(tolower(output)=="rmd",
+           "```{r echo = TRUE, eval = FALSE}",
+           "```r"),
     "#install.packages(\"devtools\", repos = \"https://cloud.r-project.org\")",
     sprintf("devtools::install_%s(\"%s/%s)",
             domain,